+++ CHANGELOG.md
@@ -192,3 +192,4 @@ was removed or the new one was added to folder
 - Nullptr check when playing sounds
 - Rounded border now has custom rounding size like rounded rectangle
 - Logger prefixes can now be modified
+- MapPos has been inlined